python/services/lipsync.py
"""
Wav2Lip Lip-Sync Service
Synchronizes lip movements with audio using Wav2Lip model
"""

import os
import logging
import torch
import cv2
import numpy as np
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from pathlib import Path

logger = logging.getLogger(__name__)

lipsync_router = APIRouter()
lipsync_model = None
lipsync_device = None


class LipSyncRequest(BaseModel):
    video_path: str  # Path to input video
    audio_path: str  # Path to audio file
    output_path: str  # Path to save synced video


def load_lipsync_model(model_dir: str):
    """Load Wav2Lip model"""
    global lipsync_model, lipsync_device
    
    try:
        checkpoint_path = os.path.join(model_dir, "wav2lip_gan.pth")
        
        if not os.path.exists(checkpoint_path):
            logger.warning(f"Wav2Lip checkpoint not found at {checkpoint_path}")
            logger.warning("Lip-sync will not be available until model is downloaded")
            return None
        
        logger.info(f"Loading Wav2Lip model from {checkpoint_path}")
        
        # Note: Wav2Lip requires additional setup and dependencies
        # This is a placeholder implementation
        # Full implementation would require:
        # - face_detection library
        # - face_alignment library
        # - Wav2Lip model architecture
        # - Video processing pipeline
        
        lipsync_device = "cuda" if torch.cuda.is_available() else "cpu"
        
        # For now, we'll mark the model as "available" if checkpoint exists
        # Actual model loading would happen here
        logger.info("Wav2Lip checkpoint found (full implementation requires additional setup)")
        
        return True  # Placeholder
    
    except Exception as e:
        logger.error(f"Error loading Wav2Lip model: {e}")
        return None


@lipsync_router.post("")
async def sync_lips(request: LipSyncRequest):
    """
    Synchronize lip movements with audio using Wav2Lip
    
    Args:
        request: LipSyncRequest with video_path, audio_path, output_path
    
    Returns:
        JSON with success status and output path
    """
    if lipsync_model is None:
        raise HTTPException(
            status_code=503,
            detail="Wav2Lip model not loaded. Please download the model first."
        )
    
    try:
        # Validate input files
        if not os.path.exists(request.video_path):
            raise HTTPException(
                status_code=400,
                detail=f"Video file not found: {request.video_path}"
            )
        
        if not os.path.exists(request.audio_path):
            raise HTTPException(
                status_code=400,
                detail=f"Audio file not found: {request.audio_path}"
            )
        
        # Create output directory if needed
        output_dir = os.path.dirname(request.output_path)
        if output_dir and not os.path.exists(output_dir):
            os.makedirs(output_dir, exist_ok=True)
        
        logger.info(f"Lip-syncing video: {request.video_path}")
        logger.info(f"Audio: {request.audio_path}")
        logger.info(f"Output: {request.output_path}")
        
        # NOTE: This is a placeholder implementation
        # Full Wav2Lip implementation requires:
        # 1. Face detection in video frames
        # 2. Face alignment and cropping
        # 3. Wav2Lip model inference
        # 4. Video reconstruction with synced lips
        # 5. Audio replacement
        
        # For MVP, we'll use a simple approach:
        # Replace audio track in video (basic implementation)
        # Full lip-sync requires the complete Wav2Lip pipeline
        
        logger.warning("Full Wav2Lip implementation requires additional dependencies")
        logger.warning("Using basic audio replacement for now")
        
        # Basic implementation: replace audio in video using ffmpeg
        # This is a placeholder - full implementation would use Wav2Lip model
        import subprocess
        
        cmd = [
            "ffmpeg",
            "-i", request.video_path,
            "-i", request.audio_path,
            "-c:v", "copy",
            "-c:a", "aac",
            "-map", "0:v:0",
            "-map", "1:a:0",
            "-shortest",
            "-y",
            request.output_path
        ]
        
        result = subprocess.run(
            cmd,
            capture_output=True,
            text=True
        )
        
        if result.returncode != 0:
            raise HTTPException(
                status_code=500,
                detail=f"FFmpeg failed: {result.stderr}"
            )
        
        if not os.path.exists(request.output_path):
            raise HTTPException(
                status_code=500,
                detail="Failed to generate output video"
            )
        
        logger.info(f"Lip-sync completed: {request.output_path}")
        
        return {
            "success": True,
            "output_path": request.output_path,
            "note": "Basic audio replacement applied. Full Wav2Lip requires additional setup."
        }
    
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Error in lip-sync: {e}")
        raise HTTPException(status_code=500, detail=f"Lip-sync failed: {str(e)}")
