python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
transformers>=4.35.0
TTS>=0.20.0
openai-whisper>=20231117
sentencepiece>=0.1.99
soundfile>=0.12.1
tqdm>=4.66.0
numpy>=1.24.0
librosa>=0.10.1
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

